agent2/requirements.txt
fastapi
uvicorn[standard]
gitpython
requests
